feat: Начал создание сетевого слоя через Swift OpenAPI Generator
- Добавил зависимости swift-openapi-generator, swift-openapi-runtime, swift-openapi-urlsession
- Подключил плагин генератора в Build Phases
- Создал файлы openapi.yaml и openapi-generator-config.yaml
- Исправил синтаксис yaml в конфиге и схеме
- Настроил servers и tags в openapi.yaml для генерации клиента
