--- Travel Schedule/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  Travel Schedule
-//
-//  Created by Ульта on 29.09.2025. //
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
